+++ ethereum/datastructures/src/main/java/tech/pegasys/teku/datastructures/blocks/SignedBeaconBlock.java
@@ -81,6 +81,15 @@ public class SignedBeaconBlock implements SimpleOffsetSerializable, SSZContainer
     return message.hash_tree_root();
   }
 
+  /**
+   * Get the state root of the BeaconBlock that is being signed.
+   *
+   * @return The hashed tree root of the {@code BeaconBlock} being signed.
+   */
+  public Bytes32 getStateRoot() {
+    return message.getState_root();
+  }
+
   @Override
   public boolean equals(final Object o) {
     if (this == o) {