+++ networking/eth2/src/test/java/tech/pegasys/teku/networking/eth2/gossip/AttestationGossipManagerTest.java
@@ -56,7 +56,7 @@ public class AttestationGossipManagerTest {
   @BeforeEach
   public void setup() {
     BeaconChainUtil.create(0, recentChainData).initializeStorage();
-    doReturn(topicChannel).when(gossipNetwork).subscribe(contains("beacon_attestation"), any());
+    doReturn(topicChannel).when(gossipNetwork).subscribe(contains("committee_index"), any());
     AttestationSubnetSubscriptions attestationSubnetSubscriptions =
         new AttestationSubnetSubscriptions(
             gossipNetwork,
@@ -78,7 +78,7 @@ public class AttestationGossipManagerTest {
     final Attestation attestation = dataStructureUtil.randomAttestation();
     setCommitteeIndex(attestation, committeeIndex);
     final Bytes serialized = gossipEncoding.encode(attestation);
-    attestationGossipManager.onNewAttestation(ValidateableAttestation.fromSingle(attestation));
+    attestationGossipManager.onNewAttestation(ValidateableAttestation.fromAttestation(attestation));
 
     verify(topicChannel).gossip(serialized);
 
@@ -86,7 +86,8 @@ public class AttestationGossipManagerTest {
     final Attestation attestation2 = dataStructureUtil.randomAttestation();
     setCommitteeIndex(attestation2, committeeIndex + ATTESTATION_SUBNET_COUNT);
     final Bytes serialized2 = gossipEncoding.encode(attestation2);
-    attestationGossipManager.onNewAttestation(ValidateableAttestation.fromSingle(attestation2));
+    attestationGossipManager.onNewAttestation(
+        ValidateableAttestation.fromAttestation(attestation2));
 
     verify(topicChannel).gossip(serialized2);
   }
@@ -100,7 +101,7 @@ public class AttestationGossipManagerTest {
     // Post new attestation
     final Attestation attestation = dataStructureUtil.randomAttestation();
     setCommitteeIndex(attestation, committeeIndex + 1);
-    attestationGossipManager.onNewAttestation(ValidateableAttestation.fromSingle(attestation));
+    attestationGossipManager.onNewAttestation(ValidateableAttestation.fromAttestation(attestation));
 
     verifyNoInteractions(topicChannel);
   }
@@ -119,7 +120,7 @@ public class AttestationGossipManagerTest {
     final Attestation attestation = dataStructureUtil.randomAttestation();
     setCommitteeIndex(attestation, dismissedIndex);
     final Bytes serialized = gossipEncoding.encode(attestation);
-    attestationGossipManager.onNewAttestation(ValidateableAttestation.fromSingle(attestation));
+    attestationGossipManager.onNewAttestation(ValidateableAttestation.fromAttestation(attestation));
 
     verify(topicChannel, never()).gossip(serialized);
 
@@ -127,7 +128,8 @@ public class AttestationGossipManagerTest {
     final Attestation attestation2 = dataStructureUtil.randomAttestation();
     setCommitteeIndex(attestation2, committeeIndex);
     final Bytes serialized2 = gossipEncoding.encode(attestation2);
-    attestationGossipManager.onNewAttestation(ValidateableAttestation.fromSingle(attestation2));
+    attestationGossipManager.onNewAttestation(
+        ValidateableAttestation.fromAttestation(attestation2));
 
     verify(topicChannel).gossip(serialized2);
   }