+++ CHANGELOG.md
@@ -7,6 +7,8 @@ we recommend most users use the latest `master` branch of Teku.
 
  - Anyone using `/node/version` should switch to use
    the new endpoint, as `/node/version` will be removed in a future release.
+ - `--metrics-host-whitelist` CLI option will be renamed `--metrics-host-allowlist` (currently both are supported)
+ - `--rest-api-host-whitelist` CLI option will be renamed `--rest-api-host-allowlist` (currently both are supported)
 
 ## 0.11.4
 
@@ -18,6 +20,7 @@ we recommend most users use the latest `master` branch of Teku.
 
 - Added `/v1/node/version` and `/v1/node/identity` REST endpoints. Anyone using `/node/version` should switch to use
 the new endpoint, as `/node/version` will be removed in a future release.
+- Added `--validators-graffiti="GRAFFITI"` command line option to allow graffiti to be used in block production.
 - Eth1 deposits now load on startup from the local database, then sync to the eth1 provider once loading is complete.
   A hidden flag has been added to disable this functionality if it causes any issues - `--Xeth1-deposits-from-storage-enabled=false`.
   Local storage requirements will increase slightly due to the need to store each deposit block